[XPU][PIR] Consolidate conv2d+bn(+add)+act XPU fusion passes
Refactor the four separate conv2d_* XPU fusion passes
(`conv2d_bn_xpu_fuse_pass`, `conv2d_bn_act_xpu_fuse_pass`,
`conv2d_bn_add_act_xpu_fuse_pass`, `depthwise_conv2d_xpu_fuse_pass`)
into a single unified `conv2d_xpu_fuse_pass`, following the design of
`fc_xpu_fuse_pass`.

Highlights:
- One pass file holding four DRR patterns with explicit benefit scores
  so longer subgraphs (bn+add+act) win over shorter ones (bn).
- Bottom-up traversal (`use_top_down_traversal=false`) so patterns
  anchored on later ops (relu) match before patterns anchored on
  earlier ops (bn).
- Shared helpers (`BuildBnFoldSubgraph`, `BuildFilterMaxSubgraph`,
  `CheckConvBnConstraints`, ComputeAttr factories) eliminate the ~50%
  duplication that existed across the original four files.
- Pattern selection via `XPU_PADDLE_CONV2D_PATTERN` env var
  (default 0xff, bits gate individual pattern groups).

Tests are consolidated into a single `test_conv2d_xpu_fuse_pass.py`
matching the unified pass name, with four `PassTest` classes covering
depthwise-only, conv+bn, conv+bn+act, and conv+bn+add+act. The
add+act test expects `pd_op.add: 1` to account for the
`bn_var + epsilon` add emitted by the BN-fold subgraph (operates on
persistable constants; can be removed by a subsequent constant-folding
pass).
