+++ paddle/fluid/inference/api/paddle_pass_builder.cc
@@ -665,11 +665,8 @@ const std::vector<std::string> kPirXpuPasses{
     "add_layernorm_xpu_fuse_pass",
     "rms_norm_xpu_fuse_pass",
     "elementwise_mul_add_xpu_fuse_pass",
-    "conv2d_bn_add_act_xpu_fuse_pass",
-    "conv2d_bn_act_xpu_fuse_pass",
-    "conv2d_bn_xpu_fuse_pass",
+    "conv2d_xpu_fuse_pass",
     "conv2d_add_xpu_fuse_pass",
-    "depthwise_conv2d_xpu_fuse_pass",
     "group_norm_silu_fuse_pass",
     "fc_xpu_fuse_pass"};
