+++ paddle/fluid/pir/transforms/passes.h
@@ -99,10 +99,7 @@ USE_PIR_PASS(add_activation_xpu_fuse_pass);
 USE_PIR_PASS(add_layernorm_xpu_fuse_pass);
 USE_PIR_PASS(rms_norm_xpu_fuse_pass);
 USE_PIR_PASS(elementwise_mul_add_xpu_fuse_pass);
-USE_PIR_PASS(conv2d_bn_xpu_fuse_pass);
-USE_PIR_PASS(conv2d_bn_act_xpu_fuse_pass);
-USE_PIR_PASS(conv2d_bn_add_act_xpu_fuse_pass);
-USE_PIR_PASS(depthwise_conv2d_xpu_fuse_pass);
+USE_PIR_PASS(conv2d_xpu_fuse_pass);
 USE_PIR_PASS(conv2d_add_xpu_fuse_pass);
 USE_PIR_PASS(fc_xpu_fuse_pass);
 #endif
