--- paddle/fluid/pir/transforms/xpu/conv2d_bn_add_act_xpu_fuse_pass.h
@@ -1,24 +1,0 @@
-// Copyright (c) 2026 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include <memory>
-#include "paddle/pir/include/pass/pass.h"
-
-namespace pir {
-
-std::unique_ptr<Pass> CreateConv2dBnAddActFuseXpuPass();
-
-}  // namespace pir
