--- test/ir/pir/fused_pass/xpu/test_conv2d_bn_act_xpu_fuse_pass.py
@@ -1,116 +1,0 @@
-# Copyright (c) 2026 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import unittest
-
-import numpy as np
-from pass_test import PassTest
-
-import paddle
-from paddle.base import core
-
-paddle.enable_static()
-
-
-class TestConv2dBnActXpuFusePattern(PassTest):
-    r"""
-    x_var   f_var
-      \       /
-         conv2d (or depthwise_conv2d)
-           |
-        BatchNorm
-           |
-          Act
-           |
-          out
-    """
-
-    def is_program_valid(self, program=None):
-        return True
-
-    def build_ir_program(
-        self, in_channels, groups, act_layer, use_global_stats
-    ):
-        with paddle.pir_utils.IrGuard():
-            main_prog = paddle.static.Program()
-            start_prog = paddle.static.Program()
-            with paddle.pir.core.program_guard(main_prog, start_prog):
-                x = paddle.static.data(
-                    name='x',
-                    shape=[2, in_channels, 16, 16],
-                    dtype='float32',
-                )
-                conv2d = paddle.nn.Conv2D(
-                    in_channels=in_channels,
-                    out_channels=in_channels if groups > 1 else 32,
-                    kernel_size=3,
-                    padding=1,
-                    groups=groups,
-                    data_format='NCHW',
-                    bias_attr=False,
-                )
-                bn = paddle.nn.BatchNorm2D(
-                    num_features=conv2d._out_channels,
-                    data_format='NCHW',
-                    use_global_stats=use_global_stats,
-                )
-                out = act_layer(bn(conv2d(x)))
-                out = paddle.assign(out)
-                self.feeds = {
-                    "x": np.random.random((2, in_channels, 16, 16)).astype(
-                        "float32"
-                    ),
-                }
-                self.fetch_list = [out]
-                return [main_prog, start_prog]
-
-    def sample_program(self):
-        # (in_channels, groups) covers both regular and depthwise convolutions.
-        conv_cfgs = [(8, 1), (8, 8)]
-        act_layers = [
-            paddle.nn.ReLU(),
-            paddle.nn.Swish(),
-            paddle.nn.Hardswish(),
-        ]
-        # use_global_stats=True  -> pd_op.batch_norm
-        # use_global_stats=False -> pd_op.batch_norm_ (inplace)
-        for use_global_stats in (True, False):
-            for in_channels, groups in conv_cfgs:
-                for act_layer in act_layers:
-                    yield (
-                        self.build_ir_program(
-                            in_channels, groups, act_layer, use_global_stats
-                        ),
-                        False,
-                    )
-
-    def test_check_output(self):
-        self.check_pass_correct(atol=2e-3, rtol=2e-3)
-
-    def setUp(self):
-        if core.is_compiled_with_xpu():
-            self.places.append(paddle.device.XPUPlace(0))
-            self.pass_attr_list = [{'conv2d_bn_act_xpu_fuse_pass': {}}]
-            self.valid_op_map = {
-                "pd_op.conv2d_xpu": 1,
-                "pd_op.batch_norm_": 0,
-                "pd_op.batch_norm": 0,
-                "pd_op.relu": 0,
-                "pd_op.swish": 0,
-                "pd_op.hardswish": 0,
-            }
-
-
-if __name__ == "__main__":
-    unittest.main()
